refactor(dkg): align DKGSessionInfo structure with genesis contract
- Restructure DKGSessionInfo to contain nested DKGSessionMetadata
- Change ConfigV2Data threshold types from uint64 to uint128 for FixedPoint64 compatibility
- Add getDKGState() view function returning 4 fields for debugging
- Update DKGStartEvent to emit DKGSessionMetadata instead of full session info
- Update spec documentation to reflect all structural changes

The DKGSessionInfo now contains:
- metadata: DKGSessionMetadata (dealerEpoch, randomnessConfig, validator sets)
- startTimeUs: uint64
- transcript: bytes

This keeps the structure consistent with gravity-genesis-contract.

+++ src/runtime/RandomnessConfig.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.30;
 
-import { SystemAddresses } from "../foundation/SystemAddresses.sol";
-import { requireAllowed } from "../foundation/SystemAccessControl.sol";
-import { Errors } from "../foundation/Errors.sol";
+import {SystemAddresses} from "../foundation/SystemAddresses.sol";
+import {requireAllowed} from "../foundation/SystemAccessControl.sol";
+import {Errors} from "../foundation/Errors.sol";
 
 /// @title RandomnessConfig
 /// @author Gravity Team
@@ -29,12 +29,13 @@ contract RandomnessConfig {
     struct ConfigV2Data {
         /// @notice Minimum stake ratio to keep randomness secret
         /// @dev Any subset with power/total <= this cannot reconstruct
-        uint64 secrecyThreshold;
+        /// @dev Uses fixed-point representation (value / 2^64), stored as uint128
+        uint128 secrecyThreshold;
         /// @notice Minimum stake ratio to reveal randomness
         /// @dev Any subset with power/total > this can reconstruct
-        uint64 reconstructionThreshold;
+        uint128 reconstructionThreshold;
         /// @notice Threshold for optimistic fast path execution
-        uint64 fastPathSecrecyThreshold;
+        uint128 fastPathSecrecyThreshold;
     }
 
     /// @notice Complete randomness configuration
@@ -68,7 +69,10 @@ contract RandomnessConfig {
     /// @notice Emitted when configuration is applied at epoch boundary
     /// @param oldVariant Previous configuration variant
     /// @param newVariant New configuration variant
-    event RandomnessConfigUpdated(ConfigVariant indexed oldVariant, ConfigVariant indexed newVariant);
+    event RandomnessConfigUpdated(
+        ConfigVariant indexed oldVariant,
+        ConfigVariant indexed newVariant
+    );
 
     /// @notice Emitted when pending configuration is set by governance
     /// @param variant The pending configuration variant
@@ -84,9 +88,7 @@ contract RandomnessConfig {
     /// @notice Initialize the randomness configuration
     /// @dev Can only be called once by GENESIS
     /// @param config Initial configuration
-    function initialize(
-        RandomnessConfigData calldata config
-    ) external {
+    function initialize(RandomnessConfigData calldata config) external {
         requireAllowed(SystemAddresses.GENESIS);
 
         if (_initialized) {
@@ -114,7 +116,11 @@ contract RandomnessConfig {
 
     /// @notice Get current active configuration
     /// @return Current configuration data
-    function getCurrentConfig() external view returns (RandomnessConfigData memory) {
+    function getCurrentConfig()
+        external
+        view
+        returns (RandomnessConfigData memory)
+    {
         _requireInitialized();
         return _currentConfig;
     }
@@ -122,7 +128,11 @@ contract RandomnessConfig {
     /// @notice Get pending configuration if any
     /// @return hasPending Whether a pending config exists
     /// @return config The pending configuration (only valid if hasPending is true)
-    function getPendingConfig() external view returns (bool hasPending, RandomnessConfigData memory config) {
+    function getPendingConfig()
+        external
+        view
+        returns (bool hasPending, RandomnessConfigData memory config)
+    {
         _requireInitialized();
         return (hasPendingConfig, _pendingConfig);
     }
@@ -140,9 +150,7 @@ contract RandomnessConfig {
     /// @notice Set configuration for next epoch
     /// @dev Only callable by GOVERNANCE. Config will be applied at epoch boundary.
     /// @param newConfig New configuration to apply at next epoch
-    function setForNextEpoch(
-        RandomnessConfigData calldata newConfig
-    ) external {
+    function setForNextEpoch(RandomnessConfigData calldata newConfig) external {
         requireAllowed(SystemAddresses.GOVERNANCE);
         _requireInitialized();
 
@@ -188,7 +196,11 @@ contract RandomnessConfig {
     /// @notice Create an Off configuration (randomness disabled)
     /// @return Configuration with Off variant
     function newOff() external pure returns (RandomnessConfigData memory) {
-        return RandomnessConfigData({ variant: ConfigVariant.Off, configV2: ConfigV2Data(0, 0, 0) });
+        return
+            RandomnessConfigData({
+                variant: ConfigVariant.Off,
+                configV2: ConfigV2Data(0, 0, 0)
+            });
     }
 
     /// @notice Create a V2 configuration with thresholds
@@ -197,18 +209,19 @@ contract RandomnessConfig {
     /// @param fastPathSecrecyThreshold Threshold for fast path (fixed-point)
     /// @return Configuration with V2 variant
     function newV2(
-        uint64 secrecyThreshold,
-        uint64 reconstructionThreshold,
-        uint64 fastPathSecrecyThreshold
+        uint128 secrecyThreshold,
+        uint128 reconstructionThreshold,
+        uint128 fastPathSecrecyThreshold
     ) external pure returns (RandomnessConfigData memory) {
-        return RandomnessConfigData({
-            variant: ConfigVariant.V2,
-            configV2: ConfigV2Data({
-                secrecyThreshold: secrecyThreshold,
-                reconstructionThreshold: reconstructionThreshold,
-                fastPathSecrecyThreshold: fastPathSecrecyThreshold
-            })
-        });
+        return
+            RandomnessConfigData({
+                variant: ConfigVariant.V2,
+                configV2: ConfigV2Data({
+                    secrecyThreshold: secrecyThreshold,
+                    reconstructionThreshold: reconstructionThreshold,
+                    fastPathSecrecyThreshold: fastPathSecrecyThreshold
+                })
+            });
     }
 
     // ========================================================================
@@ -222,8 +235,13 @@ contract RandomnessConfig {
     ) internal pure {
         if (config.variant == ConfigVariant.V2) {
             // For V2, reconstruction threshold must be >= secrecy threshold
-            if (config.configV2.reconstructionThreshold < config.configV2.secrecyThreshold) {
-                revert Errors.InvalidRandomnessConfig("reconstruction must be >= secrecy");
+            if (
+                config.configV2.reconstructionThreshold <
+                config.configV2.secrecyThreshold
+            ) {
+                revert Errors.InvalidRandomnessConfig(
+                    "reconstruction must be >= secrecy"
+                );
             }
         }
         // ConfigVariant.Off requires no validation
@@ -236,4 +254,3 @@ contract RandomnessConfig {
         }
     }
 }
-
